server/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# PDF Processing
pdfplumber==0.10.3
PyMuPDF==1.23.8
pandas==2.1.4
tabulate==0.9.0  # For markdown table conversion

# Unstructured (Advanced PDF Extraction)
unstructured[pdf]==0.12.4
unstructured-inference==0.7.23
pdf2image==1.16.3
pdfminer.six==20221105
pikepdf==8.10.1

# LangChain
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.1

# Vector Store
chromadb==0.4.22
numpy==1.24.3

# Image processing (keep Pillow for PDF image extraction)
Pillow>=10.2.0,<12.0  # For image processing in PDFs

# Groq API
groq==0.4.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Authentication & Database
motor==3.3.2
pymongo==4.6.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.26.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
